Split out our embedded assets to a `warp_assets` crate.
