+++ crates/warp_assets/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "warp_assets"
+version = "0.0.0"
+edition = "2024"
+authors.workspace = true
+license.workspace = true
+publish.workspace = true
+
+[features]
+standalone = []
+
+[dependencies]
+anyhow.workspace = true
+rust-embed.workspace = true
+warpui_core.workspace = true
+
+[target.'cfg(target_family = "wasm")'.dependencies]
+# Make sure we embed the data into the wasm binary even when running in debug
+# mode, as it cannot be read directly from the filesystem.
+rust-embed = { workspace = true, features = ["debug-embed"] }
